rename files to reflect whether it contains abstract class or not + address PR comments
