+++ examples/demo/lightning_mixed_regression.py
@@ -8,7 +8,7 @@ from pyrelational.informativeness import (
     regression_least_confidence,
     representative_sampling,
 )
-from pyrelational.strategies.strategy import Strategy
+from pyrelational.strategies.abstract_strategy import Strategy
 
 
 class MixedStrategy(Strategy):
