+++ pyrelational/models/__init__.py
@@ -1,7 +1,7 @@
+from pyrelational.models.abstract_model_manager import ModelManager
 from pyrelational.models.ensemble_model import EnsembleManager, LightningEnsembleModel
 from pyrelational.models.lightning_model import LightningModel
 from pyrelational.models.mcdropout_model import (
     LightningMCDropoutModel,
     MCDropoutManager,
 )
-from pyrelational.models.model_manager import ModelManager
