+++ pyrelational/models/mcdropout_model.py
@@ -7,8 +7,8 @@ from pytorch_lightning import LightningModule
 from torch.nn.modules import Module
 from torch.utils.data import DataLoader
 
+from .abstract_model_manager import ModelManager
 from .lightning_model import LightningModel
-from .model_manager import ModelManager
 from .model_utils import _determine_device
 
 logger = logging.getLogger()
