+++ pyrelational/strategies/__init__.py
@@ -1,4 +1,4 @@
 import pyrelational.strategies.classification
 import pyrelational.strategies.regression
 import pyrelational.strategies.task_agnostic
-from pyrelational.strategies.strategy import Strategy
+from pyrelational.strategies.abstract_strategy import Strategy
