+++ pyrelational/strategies/abstract_strategy.py
