+++ pyrelational/strategies/regression/abstract_regression_strategy.py
@@ -1,10 +1,11 @@
 from abc import ABC
+from typing import List
 
 import torch
 
 from pyrelational.data import DataManager
 from pyrelational.models import ModelManager
-from pyrelational.strategies.strategy import Strategy
+from pyrelational.strategies.abstract_strategy import Strategy
 
 
 class RegressionStrategy(Strategy, ABC):
@@ -15,7 +16,7 @@ class RegressionStrategy(Strategy, ABC):
         super(RegressionStrategy, self).__init__()
         self.scoring_fn = NotImplementedError
 
-    def __call__(self, num_annotate: int, data_manager: DataManager, model: ModelManager):
+    def __call__(self, num_annotate: int, data_manager: DataManager, model: ModelManager) -> List[int]:
         output = self.train_and_infer(data_manager=data_manager, model=model)
         scores = self.scoring_fn(x=output)
         ixs = torch.argsort(scores, descending=True).tolist()
